+++ packages/server/src/utils/sentryDsnToReportUri.ts
@@ -1,3 +1,5 @@
+import { URL } from 'url';
+
 export default function sentryDsnToReportUri(dsn: string): string {
   if (!dsn) {
     return null;