+++ packages/utils/src/blockUtils.ts
@@ -1,6 +1,6 @@
-import type { Block } from '@appsemble/types';
+import type { BlockDefinition } from '@appsemble/types';
 
-export type IdentifiableBlock = Pick<Block, 'type' | 'version'>;
+export type IdentifiableBlock = Pick<BlockDefinition, 'type' | 'version'>;
 
 /**
  * Normalize a block name by prefixing it with `@appsemble` is necessary.